[types] - refactor: move content fragment logic to a dedicated module
 - Extract ContentFragment related types and functions from the conversation module
 - Create a new module content_fragment to manage them separately
 - Update import statements in internal and public API handlers to refer to the new module location
 - Prepare for removal of deprecated content fragment types with a TODO in the new module
